Model: Refactor code

+++ config.py
@@ -4,7 +4,7 @@ import tensorflow as tf
 
 PATH_FILE = os.path.join(os.path.split(__file__)[0],
                          'quantazation/pts_in_hull.npy')
-CENTERS = tf.constant(np.load(PATH_FILE))
+CENTERS = tf.constant(np.load(PATH_FILE), dtype='float64')
 T = 0.38
 EPSILON = 1e-6
 NUM_CLASSES_Q = 313
